projectiles and collision fixes
Projectiles: projectiles is now its own class, game now has two vectors of projectiles (one for aliens and one for player) to remove the need for a type check.
Collision: collision is now handled with raylibs built in functions for collision which both reduces the amount of code and probably runs faster.